fix(stats): 全面审计并修复 /proc/firewall/stats 的 4 个真实 Bug
[Bug 1] 重复封禁时 total_ban_count/ban_count 过度计数
  __do_ban_ip_ipv4/ipv6 在"已存在且仍有效"路径错误地既无操作
  又返回 0,导致上层 __do_ban_ip 盲目 atomic_inc,每次重复 ban 都
  污染统计。修复:内层新增 -EEXIST 返回值明确区分 no-op 与
  new insert,刷新过期条目同样不再计入任一计数器。

[Bug 2] IPv6 封禁表 hash_add_rcu 桶错位(链式发现)
  __do_ban_ip_ipv6 使用 hash_add_rcu(table, node, bkt6),
  但 bkt6 已是桶索引,hash_add_rcu 会以其为 key 重新 hash_min
  落到错误桶。修复:改用 hlist_add_head_rcu 直写预计算桶。

[Bug 3] state-persist.c IPv6 恢复路径同源 Bug
  与 Bug 2 完全相同的误用,模块重启后 IPv6 ban 恢复错位。

[Bug 4] whitelist.c IPv6 桶错位(白名单保护可能失效)
  hash_wl_ipv6 结果作 hash_add_rcu 的 key,二次哈希落到错误桶。
  后果:(1) 重复 add 检查失效(5+ 重复 ::1);(2) netfilter 热
  路径查找 miss(白名单保护可能失效)。

改进:
  - 移除 fw_flush_cpu_stats() 冗余调用
  - cleanup_last_bucket_* 加 READ_ONCE/WRITE_ONCE 屏障
  - atomic_t 改用显式 (unsigned int) 强转
  - stats_show 补充 packets_* 统计范围注释
  - 统计不变量入代码注释

运行时守护:
  - cleanup_expired_bans 末尾加 WARN_ON_ONCE ±4096 容差检测
  - 任何计数漂移超阈值即打印 backtrace + delta

API 一致性收敛:
  - 6 处 hash_add_rcu 全部替换为 hlist_add_head_rcu 直写预计算桶
  - 彻底消除 hash_add_rcu(table, node, KEY) API 误用面
  - IPv4 路径此前"巧合正确",统一后杜绝未来重构风险

测试:
  - 03.5: IPv4 重复 ban 3 次仅 +1
  - 03.6: IPv6 重复 ban 3 次仅 +1(覆盖 __do_ban_ip_ipv6)
  - 04.3: whitelist_rejects 计数验证
  - 04.2/12.5: 白名单满环境优雅跳过
  - test_framework.sh: grep -qF 固定字符串匹配(防 IPv6 展开形式误命中)

文档:
  - 4 份 + CHANGELOG 完整同步
  - 守恒律契约入代码: total_bans == current_bans + total_unbans + cleanup_expired_total

验证:
  - 84/84 测试全过(75 普通 + 5 压力 + 4 并发,0 失败 0 警告 0 跳过)
  - 守恒律实测成立
  - 压力测试下 WARN_ON_ONCE 不误报

+++ src/kernel-module/procfs.c
@@ -896,21 +896,28 @@ static const struct proc_ops config_fops = {
 
 /*
  * stats_show - 显示防火墙统计信息
+ *
+ * 说明：
+ * - 以下计数器仅反映经过 ban 检查的包(分片 / 非法源 IP 不计入)。
+ * - atomic_t 为有符号整型,通过显式强转避免有符号/无符号格式符误用。
  */
 static int stats_show(struct seq_file *m, void *v) {
   struct firewall_info *fw = &fw_info;
 
-  seq_printf(m, "total_bans %u\n", atomic_read(&fw->total_ban_count));
-  seq_printf(m, "total_unbans %u\n", atomic_read(&fw->total_unban_count));
-  seq_printf(m, "whitelist_rejects %u\n", atomic_read(&fw->whitelist_reject_count));
-  seq_printf(m, "ban_table_full_rejects %u\n", atomic_read(&fw->ban_table_full_count));
-  seq_printf(m, "alloc_failures %u\n", atomic_read(&fw->alloc_failure_count));
+  seq_printf(m, "total_bans %u\n", (unsigned int)atomic_read(&fw->total_ban_count));
+  seq_printf(m, "total_unbans %u\n", (unsigned int)atomic_read(&fw->total_unban_count));
+  seq_printf(m, "whitelist_rejects %u\n",
+             (unsigned int)atomic_read(&fw->whitelist_reject_count));
+  seq_printf(m, "ban_table_full_rejects %u\n",
+             (unsigned int)atomic_read(&fw->ban_table_full_count));
+  seq_printf(m, "alloc_failures %u\n", (unsigned int)atomic_read(&fw->alloc_failure_count));
   seq_printf(m, "packets_dropped %llu\n",
              (unsigned long long)atomic64_read(&fw->packets_dropped));
   seq_printf(m, "packets_accepted %llu\n",
              (unsigned long long)atomic64_read(&fw->packets_accepted));
-  seq_printf(m, "cleanup_cycles %u\n", atomic_read(&fw->cleanup_cycles));
-  seq_printf(m, "cleanup_expired_total %u\n", atomic_read(&fw->cleanup_expired_total));
+  seq_printf(m, "cleanup_cycles %u\n", (unsigned int)atomic_read(&fw->cleanup_cycles));
+  seq_printf(m, "cleanup_expired_total %u\n",
+             (unsigned int)atomic_read(&fw->cleanup_expired_total));
   seq_printf(m, "current_bans %d\n", atomic_read(&fw->ban_count));
   seq_printf(m, "current_whitelist %d\n", atomic_read(&fw->whitelist_count));
   {
